Refactor to use double in route cost

+++ TravellingSalespersonProj/TravellingSalespersonProj/EvolutionaryAlgorithms/WheelSegment.cs
@@ -5,9 +5,9 @@ namespace TravellingSalespersonProj.EvolutionaryAlgorithms
     public class WheelSegment
     {
         public Route Route { get; set; }
-        public float LowerBound { get; set; }
+        public double LowerBound { get; set; }
 
-        public WheelSegment(Route route, float lowerBound)
+        public WheelSegment(Route route, double lowerBound)
         {
             Route = route;
             LowerBound = lowerBound;
