+++ TravellingSalespersonProj/TravellingSalespersonProj/GenericTSP/Route.cs
@@ -7,9 +7,9 @@ namespace TravellingSalespersonProj.LocalSearchTutorial
     public class Route
     {
         public int[] RouteIds { get; set; }
-        public float RouteCost { get; set; }
+        public double RouteCost { get; set; }
 
-        public Route(int[] routeIds, float routeCost)
+        public Route(int[] routeIds, double routeCost)
         {
             RouteIds = routeIds;
             RouteCost = routeCost;
